Place code in the datasets/map.py, remove from core
