+++ gammapy/datasets/tests/test_datasets.py
@@ -2,7 +2,6 @@
 import pytest
 from numpy.testing import assert_allclose
 from gammapy.datasets import Datasets, SpectrumDatasetOnOff
-from gammapy.maps import MapAxis, RegionGeom
 from gammapy.modeling.tests.test_fit import MyDataset
 from gammapy.utils.testing import requires_data
 
@@ -88,11 +87,3 @@ def test_datasets_info_table():
     table = datasets_hess.info_table(cumulative=True)
     assert table["name"][0] == "stacked"
     assert table["name"][1] == "stacked"
-
-
-def test_stat_sum():
-    axis = MapAxis.from_energy_bounds(0.1, 10, 5, unit="TeV")
-    geom = RegionGeom.create(None, axes=[axis])
-    dataset = SpectrumDatasetOnOff.create(geom)
-    dataset.counts_off = None
-    return dataset.stat_sum()
